requirements.txt
fabric-cicd>=0.3.1
azure-identity>=1.17.0
